Revert "Host: Add GetTopLevelWindowInfo()"
This reverts commit 8d7aea5e19859ed483699cc4a5dbd47165c7be8b.

This caused a deadlock on Linux when clicking Debug > CPU Debugger
when the "Inhibit Screensaver" preference is on.

The main window thread blocked waiting for the emulation thread to
pause. The emulation thread, when it pauses, wants to turn the
screensaver back on, which blocks on the main window thread
waiting for it to return its window info.